Exercise 6 complete
